requirements.txt
flask>=3.1.1,<4.0
langchain-community>=0.3.25,<1.0
langchain-openai>=0.1.10,<1.0
openai>=1.88.0,<2.0
faiss-cpu>=1.11.0,<2.0
python-dotenv>=1.1.0,<2.0
tiktoken>=0.6.0,<1.0